+++ app/src/main/java/com/example/features/weather/model/HoursWeather.kt
@@ -2,6 +2,6 @@ package com.example.features.weather.model
 
 data class HoursWeather(
     val time: String,
-    val temp: Float,
+    val temp: Double,
     val icon: String
 )
