+++ app/src/main/java/com/example/features/weather/model/WeatherDetailedDay.kt
@@ -2,12 +2,12 @@ package com.example.features.weather.model
 
 data class WeatherDetailedDay(
     val city: String,
-    val temp: Float,
+    val temp: Double,
     val descriptionWeather: String,
-    val maxTemp: Float,
-    val minTemp: Float,
-    val feelingTemp: Float,
+    val maxTemp: Double,
+    val minTemp: Double,
+    val feelingTemp: Double,
 )
 
 val emptyWeatherDetailedDay =
-    WeatherDetailedDay("",0.0f, "" , 0.0f,0.0f,0.0f)
+    WeatherDetailedDay("",0.0, "" , 0.0,0.0,0.0)
